refactor(text-sync): split document state and tests (#8197)
